requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.16.0
sqlparse==0.5.3
gunicorn==21.2.0
python-dotenv==1.1.0
djangorestframework_simplejwt==5.5.0
requests==2.32.3
idna==3.10
PyJWT==2.9.0
urllib3==2.4.0
